requirements.txt
torch
pandas
numpy
faiss-cpu
sentence-transformers
datasets
tqdm
wandb
datasets